Separate files, reorganized, started unit tests on Devices
