+++ CommunicationPort.py
@@ -1,8 +1,8 @@
 import serial
 import re
-from threading import Thread, RLock
 import time
 import random
+from threading import Thread, RLock
 
 class CommunicationReadTimeout(serial.SerialException):
     pass
@@ -12,29 +12,35 @@ class CommunicationReadNoMatch(Exception):
 
 class CommunicationPort:
     """CommunicationPort class with basic application-level protocol 
-    functions to write strings and read strings"""
+    functions to write strings and read strings, and abstract away
+    the details of the communication.
+
+    Three strategies to initialize the CommunicationPort:
+    1. with a bsdPath/port name (i.e. "COM1" or "/dev/cu.serial")
+    2. with an instance of SerialPort() that will support the same
+       functions as pyserial.Serial() (open, close, read, write, readline)
+    """
     
-    def __init__(self, bsdPath=None, port = None, vendorId=None,
-                 productId=None, serialNumber=None):
+    def __init__(self, bsdPath=None, port = None):
         self.bsdPath = bsdPath
-        self.vendorId = vendorId
-        self.productId = productId
-        self.serialNumber = serialNumber
+        self.port = port # direct port , must be closed.
+
         self.portLock = RLock()
         self.transactionLock = RLock()
-        self.port = port # direct port assignment
 
     def open(self):
         if self.port is None:
             self.port = serial.Serial(self.bsdPath, 19200, timeout=0.3)
+        else:
+            self.port.open()
 
     def close(self):
         self.port.close()
 
-    def bytesAvailable(self):
+    def bytesAvailable(self) -> int:
         return self.port.in_waiting
 
-    def readData(self, length):
+    def readData(self, length) -> bytearray:
         with self.portLock:
             data = self.port.read(length)
             if len(data) != length:
@@ -42,7 +48,7 @@ class CommunicationPort:
 
         return data
 
-    def writeData(self, data):
+    def writeData(self, data) -> int:
         with self.portLock:
             nBytesWritten = self.port.write(data)
             if nBytesWritten != len(data):
@@ -51,7 +57,7 @@ class CommunicationPort:
 
         return nBytesWritten
 
-    def readString(self):      
+    def readString(self) -> str:      
         with self.portLock:
             byte = None
             data = bytearray(0)
@@ -65,9 +71,9 @@ class CommunicationPort:
 
         return string
 
-    def writeString(self, string):
+    def writeString(self, string) -> int:
+        nBytes = 0
         with self.portLock:
-            nBytes = 0
             data = bytearray(string, "utf-8")
             nBytes = self.writeData(data)
 
@@ -95,6 +101,7 @@ class CommunicationPort:
         with self.transactionLock:
             self.writeString(string)
             reply = self.readString()
+
             match = re.search(replyPattern, reply)
 
             if match is not None:
@@ -103,36 +110,3 @@ class CommunicationPort:
                 raise CommunicationReadNoMatch("Unable to match pattern:'{0}' in reply:'{1}'".format(replyPattern, reply))
 
 
-class DebugEchoCommunicationPort(CommunicationPort):
-    def __init__(self, delay=0):
-        self.buffer = bytearray()
-        self.delay = delay
-        super(DebugEchoCommunicationPort, self).__init__()
-
-    def open(self):
-        return
-
-    def close(self):
-        return
-
-    def bytesAvailable(self):
-        return len(self.buffer)
-
-    def readData(self, length):
-        with self.portLock:
-            time.sleep(self.delay*random.random())
-            data = bytearray()
-            for i in range(0, length):
-                if len(self.buffer) > 0:
-                    byte = self.buffer.pop(0)
-                    data.append(byte)
-                else:
-                    raise CommunicationReadTimeout("Unable to read data")
-
-        return data
-
-    def writeData(self, data):
-        with self.portLock:
-            self.buffer.extend(data)
-
-        return len(data)
