+++ DebugEchoCommunicationPort.py
@@ -1,0 +1,35 @@
+from CommunicationPort import *
+
+class DebugEchoCommunicationPort(CommunicationPort):
+    def __init__(self, delay=0):
+        self.buffer = bytearray()
+        self.delay = delay
+        super(DebugEchoCommunicationPort, self).__init__()
+
+    def open(self):
+        return
+
+    def close(self):
+        return
+
+    def bytesAvailable(self):
+        return len(self.buffer)
+
+    def readData(self, length):
+        with self.portLock:
+            time.sleep(self.delay*random.random())
+            data = bytearray()
+            for i in range(0, length):
+                if len(self.buffer) > 0:
+                    byte = self.buffer.pop(0)
+                    data.append(byte)
+                else:
+                    raise CommunicationReadTimeout("Unable to read data")
+
+        return data
+
+    def writeData(self, data):
+        with self.portLock:
+            self.buffer.extend(data)
+
+        return len(data)
